refactor: 안보이는 로그들 삭제
