+++ reflective-client/src/types/BlockEditor/Chart.ts
@@ -13,3 +13,11 @@ export interface ChartEditorProps {
   isEditing: boolean;
   setIsEditing: React.Dispatch<React.SetStateAction<boolean>>;
 }
+export interface ChartData {
+  labels: string[];
+  datasets: {
+    label: string;
+    data: number[];
+    borderColor?: string;
+  }[];
+}
